+++ front/scripts/remove_draft_agent_configurations.ts
@@ -14,8 +14,8 @@ import { Mention } from "@app/lib/models/assistant/conversation";
 import { Workspace } from "@app/lib/models/workspace";
 import { renderLightWorkspaceType } from "@app/lib/workspace";
 import type { Logger } from "@app/logger/logger";
-import { makeScript, runOnAllWorkspaces } from "@app/scripts/helpers";
-
+import { makeScript } from "@app/scripts/helpers";
+import { runOnAllWorkspaces } from "@app/scripts/workspace_helpers";
 async function deleteRetrievalConfigurationForAgent(agent: AgentConfiguration) {
   const retrievalConfigurations = await AgentRetrievalConfiguration.findAll({
     where: {
